PDI-11201 - Modified processList, processRecord in NoAuthMongoClientWrapper to allow for null values, added fields to unit test json to validate, deleted dead code
